+++ app/app/bookmarks/index.tsx
@@ -1,5 +1,5 @@
-import { Image, StyleSheet, Text, View } from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
+import BookmarkCard from "@src/components/bookmark-card";
+import { View } from "react-native";
 
 const Index = () => {
     const cards = [
@@ -18,58 +18,15 @@ const Index = () => {
     ];
 
     return (
-        <SafeAreaView>
+        <View>
             {cards.map((card, index) => (
-                <View style={styles.card} key={index}>
-                    <View style={styles.cardContent}>
-                        <View style={styles.cardText}>
-                            <Text style={styles.title}>{card.title}</Text>
-                            <Text style={styles.description}>
-                                {card.description}
-                            </Text>
-                        </View>
-                        <Image
-                            source={{ uri: card.imageSource }}
-                            style={styles.cardImage}
-                        />
-                    </View>
-                </View>
+                <BookmarkCard
+                    key={index}
+                    card={card}
+                />
             ))}
-        </SafeAreaView>
+        </View>
     );
 };
-
-const styles = StyleSheet.create({
-    card: {
-        backgroundColor: "#FFFFFF",
-        padding: 16,
-        marginVertical: 1,
-        shadowColor: "#000000",
-        shadowOffset: { width: 0, height: 2 },
-        shadowOpacity: 0.1,
-        shadowRadius: 4,
-    },
-    cardContent: {
-        flexDirection: "row",
-    },
-    cardText: {
-        flex: 1,
-        marginRight: 16,
-    },
-    cardImage: {
-        width: 100,
-        height: 100,
-        borderRadius: 5,
-    },
-    title: {
-        fontSize: 18,
-        fontWeight: "bold",
-        marginBottom: 8,
-    },
-    description: {
-        fontSize: 16,
-        color: "#808080",
-    },
-});
 
 export default Index;
